refactor(tool): Move output path related structs, ... into tool::path module
