+++ iai-callgrind-runner/src/runner/lib_bench.rs
@@ -12,7 +12,8 @@ use super::format::{LibraryBenchmarkHeader, OutputFormat};
 use super::meta::Metadata;
 use super::summary::{BaselineKind, BaselineName, BenchmarkKind, BenchmarkSummary, SummaryOutput};
 use super::tool::config::ToolConfigs;
-use super::tool::{RunOptions, ToolOutputPath, ToolOutputPathKind};
+use super::tool::path::{ToolOutputPath, ToolOutputPathKind};
+use super::tool::RunOptions;
 use crate::api::{
     EntryPoint, LibraryBenchmarkBench, LibraryBenchmarkConfig, LibraryBenchmarkGroups, RawArgs,
     ValgrindTool,
