+++ iai-callgrind-runner/src/runner/tool/generic_parser.rs
@@ -6,7 +6,7 @@ use anyhow::{Context, Result};
 
 use super::logfile_parser::{parse_header, EMPTY_LINE_RE, STRIP_PREFIX_RE};
 use super::parser::{Parser, ParserOutput};
-use super::ToolOutputPath;
+use super::path::ToolOutputPath;
 use crate::runner::summary::ToolMetrics;
 
 #[derive(Debug, PartialEq, Eq, Clone, Copy)]
