+++ iai-callgrind-runner/tests/common/mod.rs
@@ -8,7 +8,7 @@ use std::process::{Command, Output};
 use anyhow::Result;
 use iai_callgrind_runner::api::ValgrindTool;
 use iai_callgrind_runner::runner::summary::BaselineKind;
-use iai_callgrind_runner::runner::tool::{ToolOutputPath, ToolOutputPathKind};
+use iai_callgrind_runner::runner::tool::path::{ToolOutputPath, ToolOutputPathKind};
 use pretty_assertions::assert_eq;
 use serde::{Deserialize, Serialize};
