+++ iai-callgrind-runner/tests/test_tool/test_parse_logfile_header.rs
@@ -4,7 +4,6 @@ use std::io::{BufRead, BufReader};
 use iai_callgrind_runner::api::ValgrindTool;
 use iai_callgrind_runner::runner::tool::logfile_parser::parse_header;
 use iai_callgrind_runner::runner::tool::parser::Header;
-use iai_callgrind_runner::runner::tool::ToolOutputPathKind;
 use rstest::rstest;
 
 use crate::common::Fixtures;
@@ -44,6 +43,8 @@ fn expected_header(command: &str, pid: i32, parent_pid: Option<i32>, desc: Vec<S
     )
 )]
 fn test_parse_logfile_header(#[case] name: &str, #[case] expected_header: Header) {
+    use iai_callgrind_runner::runner::tool::path::ToolOutputPathKind;
+
     let tool_output_path =
         Fixtures::get_tool_output_path("drd", ValgrindTool::DRD, ToolOutputPathKind::Log, name);
     let mut logfile_headers = vec![];
